+++ src/main/java/uz/mediasolutions/barterlybackend/service/admin/impl/CityServiceImpl.java
@@ -19,6 +19,9 @@ import uz.mediasolutions.barterlybackend.repository.RegionRepository;
 import uz.mediasolutions.barterlybackend.service.admin.abs.CityService;
 import uz.mediasolutions.barterlybackend.utills.constants.Rest;
 
+import javax.swing.text.html.Option;
+import java.util.Optional;
+
 @Service
 @RequiredArgsConstructor
 public class CityServiceImpl implements CityService {
@@ -27,12 +30,14 @@ public class CityServiceImpl implements CityService {
     private final CityMapper cityMapper;
     private final RegionRepository regionRepository;
 
+
     @Override
     public ResponseEntity<Page<?>> getAll(String lang, String search, Long regionId, int page, int size) {
         Pageable pageable = PageRequest.of(page, size);
         Page<CityDTO> cities = cityRepository.findAllCustom(lang, search, regionId, pageable);
         return ResponseEntity.ok(cities);
     }
+
 
     @Override
     public ResponseEntity<?> getById(String lang, Long id) {
@@ -42,6 +47,7 @@ public class CityServiceImpl implements CityService {
         return ResponseEntity.ok(city);
     }
 
+
     @Override
     public ResponseEntity<?> add(CityReqDTO dto) {
         City city = cityMapper.toEntity(dto);
@@ -49,29 +55,33 @@ public class CityServiceImpl implements CityService {
         return ResponseEntity.status(HttpStatus.CREATED).body(Rest.CREATED);
     }
 
+
     @Override
     public ResponseEntity<?> edit(CityReqDTO dto, Long id) {
         City city = cityRepository.findById(id).orElseThrow(
                 () -> RestException.restThrow("City not found", HttpStatus.NOT_FOUND)
         );
 
-        Region region = regionRepository.findById(dto.getRegionId()).orElseThrow(
-                () -> RestException.restThrow("Region not found", HttpStatus.NOT_FOUND)
-        );
-        city.setTranslations(dto.getTranslations());
-        city.setRegion(region);
+        Optional.ofNullable(dto.getTranslations()).ifPresent(city::setTranslations);
+        Optional.ofNullable(dto.getRegionId()).ifPresent(regionId -> city.setRegion(
+                regionRepository.findById(dto.getRegionId()).orElseThrow(
+                        () -> RestException.restThrow("Region not found", HttpStatus.NOT_FOUND)
+                )
+        ));
         cityRepository.save(city);
         return ResponseEntity.status(HttpStatus.ACCEPTED).body(Rest.EDITED);
     }
 
+
     @Override
     public ResponseEntity<?> delete(Long id) {
-        cityRepository.findById(id).orElseThrow(
-                () -> RestException.restThrow("City not found", HttpStatus.NOT_FOUND)
-        );
+        if (!cityRepository.existsById(id)) {
+            throw RestException.restThrow("City not found", HttpStatus.NOT_FOUND);
+        }
+
         try {
             cityRepository.deleteById(id);
-            return ResponseEntity.status(HttpStatus.ACCEPTED).body(Rest.DELETED);
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).body(Rest.DELETED);
         } catch (Exception e) {
             e.printStackTrace();
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(Rest.ERROR);
